Splitting primal_grid_structure
