+++ Common/src/geometry/primal_grid/CVertexMPI.cpp
@@ -1,0 +1,66 @@
+/*!
+ * \file CVertexMPI.cpp
+ * \brief Main classes for defining the primal grid elements
+ * \author F. Palacios
+ * \version 7.0.0 "Blackbird"
+ *
+ * SU2 Project Website: https://su2code.github.io
+ *
+ * The SU2 Project is maintained by the SU2 Foundation
+ * (http://su2foundation.org)
+ *
+ * Copyright 2012-2019, SU2 Contributors (cf. AUTHORS.md)
+ *
+ * SU2 is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * SU2 is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with SU2. If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#include "../include/geometry/primal_grid/CVertexMPI.hpp"
+
+unsigned short CVertexMPI::nFaces = 0;
+
+unsigned short CVertexMPI::nNodes = 1;
+
+unsigned short CVertexMPI::nNeighbor_Elements = 0;
+
+unsigned short CVertexMPI::VTK_Type = 1;
+
+unsigned short CVertexMPI::maxNodesFace = 0;
+
+CVertexMPI::CVertexMPI(unsigned long val_point, unsigned short val_nDim) : CPrimalGrid() {
+  unsigned short iDim;
+
+  /*--- Allocate CG coordinates ---*/
+  nDim = val_nDim;
+  Coord_CG = new su2double[nDim];
+  for (iDim = 0; iDim < nDim; iDim++) Coord_CG[iDim] = 0.0;
+
+  /*--- Allocate and define face structure of the element ---*/
+  Nodes = new unsigned long[nNodes];
+  Nodes[0] = val_point;
+
+  /*--- By default, no rotation in the solution ---*/
+  Rotation_Type = 0;
+
+}
+
+CVertexMPI::~CVertexMPI() {
+  unsigned short iFaces;
+
+    for (iFaces = 0; iFaces < nFaces; iFaces++)
+      if (Coord_FaceElems_CG[iFaces] != NULL) delete[] Coord_FaceElems_CG[iFaces];
+    if (Coord_FaceElems_CG != NULL) delete[] Coord_FaceElems_CG;
+
+}
+
+void CVertexMPI::Change_Orientation(void) { }
